+++ src/main/java/com/edgar/module/sys/service/SysMenuService.java
@@ -61,7 +61,7 @@ public interface SysMenuService {
 	 * @return 如果删除成功，返回<code>1</code>，保存失败，返回<code>0</code>或者抛出
 	 *         {@link SystemException}
 	 */
-	int deleteWithLock(@Min(1) int menuId, @NotNull long updatedTime);
+	long deleteWithLock(@Min(1) int menuId, @NotNull long updatedTime);
 
 	/**
 	 * 查询菜单的列表
